LayoutTests/imported/w3c/web-platform-tests/css/selectors/invalidation/typed-child-indexed-pseudo-classes-in-has-expected.txt


PASS Initial colors: #only_of_type
PASS Initial colors: #first_of_type
PASS Initial colors: #last_of_type
PASS Initial colors: #nth_of_type_3n_1
PASS Initial colors: #nth_of_type_3n_2
PASS Initial colors: #nth_of_type_3n
PASS Prepend span (1): #only_of_type
PASS Prepend span (1): #first_of_type
PASS Prepend span (1): #last_of_type
PASS Prepend span (1): #nth_of_type_3n_1
PASS Prepend span (1): #nth_of_type_3n_2
PASS Prepend span (1): #nth_of_type_3n
PASS Prepend #div1.green: #only_of_type
PASS Prepend #div1.green: #first_of_type
PASS Prepend #div1.green: #last_of_type
PASS Prepend #div1.green: #nth_of_type_3n_1
PASS Prepend #div1.green: #nth_of_type_3n_2
PASS Prepend #div1.green: #nth_of_type_3n
PASS Prepend span (2): #only_of_type
PASS Prepend span (2): #first_of_type
PASS Prepend span (2): #last_of_type
PASS Prepend span (2): #nth_of_type_3n_1
PASS Prepend span (2): #nth_of_type_3n_2
PASS Prepend span (2): #nth_of_type_3n
FAIL Prepend #div2.yellow: #only_of_type assert_equals: expected "rgb(128, 128, 128)" but got "rgb(255, 0, 0)"
PASS Prepend #div2.yellow: #first_of_type
PASS Prepend #div2.yellow: #last_of_type
PASS Prepend #div2.yellow: #nth_of_type_3n_1
FAIL Prepend #div2.yellow: #nth_of_type_3n_2 assert_equals: expected "rgb(0, 128, 0)" but got "rgb(128, 128, 128)"
PASS Prepend #div2.yellow: #nth_of_type_3n
FAIL Prepend span (3): #only_of_type assert_equals: expected "rgb(128, 128, 128)" but got "rgb(255, 0, 0)"
PASS Prepend span (3): #first_of_type
PASS Prepend span (3): #last_of_type
PASS Prepend span (3): #nth_of_type_3n_1
FAIL Prepend span (3): #nth_of_type_3n_2 assert_equals: expected "rgb(0, 128, 0)" but got "rgb(128, 128, 128)"
PASS Prepend span (3): #nth_of_type_3n
FAIL Prepend #div3.orange: #only_of_type assert_equals: expected "rgb(128, 128, 128)" but got "rgb(255, 0, 0)"
PASS Prepend #div3.orange: #first_of_type
PASS Prepend #div3.orange: #last_of_type
PASS Prepend #div3.orange: #nth_of_type_3n_1
FAIL Prepend #div3.orange: #nth_of_type_3n_2 assert_equals: expected "rgb(255, 255, 0)" but got "rgb(128, 128, 128)"
FAIL Prepend #div3.orange: #nth_of_type_3n assert_equals: expected "rgb(0, 128, 0)" but got "rgb(128, 128, 128)"
FAIL Prepend span (4): #only_of_type assert_equals: expected "rgb(128, 128, 128)" but got "rgb(255, 0, 0)"
PASS Prepend span (4): #first_of_type
PASS Prepend span (4): #last_of_type
PASS Prepend span (4): #nth_of_type_3n_1
FAIL Prepend span (4): #nth_of_type_3n_2 assert_equals: expected "rgb(255, 255, 0)" but got "rgb(128, 128, 128)"
FAIL Prepend span (4): #nth_of_type_3n assert_equals: expected "rgb(0, 128, 0)" but got "rgb(128, 128, 128)"
FAIL Prepend #div4: #only_of_type assert_equals: expected "rgb(128, 128, 128)" but got "rgb(255, 0, 0)"
FAIL Prepend #div4: #first_of_type assert_equals: expected "rgb(128, 128, 128)" but got "rgb(255, 165, 0)"
PASS Prepend #div4: #last_of_type
FAIL Prepend #div4: #nth_of_type_3n_1 assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 165, 0)"
FAIL Prepend #div4: #nth_of_type_3n_2 assert_equals: expected "rgb(255, 165, 0)" but got "rgb(128, 128, 128)"
FAIL Prepend #div4: #nth_of_type_3n assert_equals: expected "rgb(255, 255, 0)" but got "rgb(128, 128, 128)"
FAIL Prepend span (5): #only_of_type assert_equals: expected "rgb(128, 128, 128)" but got "rgb(255, 0, 0)"
FAIL Prepend span (5): #first_of_type assert_equals: expected "rgb(128, 128, 128)" but got "rgb(255, 165, 0)"
PASS Prepend span (5): #last_of_type
FAIL Prepend span (5): #nth_of_type_3n_1 assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 165, 0)"
FAIL Prepend span (5): #nth_of_type_3n_2 assert_equals: expected "rgb(255, 165, 0)" but got "rgb(128, 128, 128)"
FAIL Prepend span (5): #nth_of_type_3n assert_equals: expected "rgb(255, 255, 0)" but got "rgb(128, 128, 128)"
FAIL Prepend #div5: #only_of_type assert_equals: expected "rgb(128, 128, 128)" but got "rgb(255, 0, 0)"
FAIL Prepend #div5: #first_of_type assert_equals: expected "rgb(128, 128, 128)" but got "rgb(255, 165, 0)"
PASS Prepend #div5: #last_of_type
FAIL Prepend #div5: #nth_of_type_3n_1 assert_equals: expected "rgb(255, 255, 0)" but got "rgb(255, 165, 0)"
FAIL Prepend #div5: #nth_of_type_3n_2 assert_equals: expected "rgb(0, 128, 0)" but got "rgb(128, 128, 128)"
FAIL Prepend #div5: #nth_of_type_3n assert_equals: expected "rgb(255, 165, 0)" but got "rgb(128, 128, 128)"
FAIL Prepend span (6): #only_of_type assert_equals: expected "rgb(128, 128, 128)" but got "rgb(255, 0, 0)"
FAIL Prepend span (6): #first_of_type assert_equals: expected "rgb(128, 128, 128)" but got "rgb(255, 165, 0)"
PASS Prepend span (6): #last_of_type
FAIL Prepend span (6): #nth_of_type_3n_1 assert_equals: expected "rgb(255, 255, 0)" but got "rgb(255, 165, 0)"
FAIL Prepend span (6): #nth_of_type_3n_2 assert_equals: expected "rgb(0, 128, 0)" but got "rgb(128, 128, 128)"
FAIL Prepend span (6): #nth_of_type_3n assert_equals: expected "rgb(255, 165, 0)" but got "rgb(128, 128, 128)"
FAIL Remove #div1: #only_of_type assert_equals: expected "rgb(128, 128, 128)" but got "rgb(255, 0, 0)"
FAIL Remove #div1: #first_of_type assert_equals: expected "rgb(128, 128, 128)" but got "rgb(255, 165, 0)"
PASS Remove #div1: #last_of_type
FAIL Remove #div1: #nth_of_type_3n_1 assert_equals: expected "rgb(255, 255, 0)" but got "rgb(255, 165, 0)"
PASS Remove #div1: #nth_of_type_3n_2
FAIL Remove #div1: #nth_of_type_3n assert_equals: expected "rgb(255, 165, 0)" but got "rgb(128, 128, 128)"
FAIL Remove #div2: #only_of_type assert_equals: expected "rgb(128, 128, 128)" but got "rgb(255, 0, 0)"
FAIL Remove #div2: #first_of_type assert_equals: expected "rgb(128, 128, 128)" but got "rgb(255, 165, 0)"
PASS Remove #div2: #last_of_type
PASS Remove #div2: #nth_of_type_3n_1
PASS Remove #div2: #nth_of_type_3n_2
FAIL Remove #div2: #nth_of_type_3n assert_equals: expected "rgb(255, 165, 0)" but got "rgb(128, 128, 128)"
FAIL Remove #div3: #only_of_type assert_equals: expected "rgb(128, 128, 128)" but got "rgb(255, 0, 0)"
FAIL Remove #div3: #first_of_type assert_equals: expected "rgb(128, 128, 128)" but got "rgb(255, 165, 0)"
PASS Remove #div3: #last_of_type
PASS Remove #div3: #nth_of_type_3n_1
PASS Remove #div3: #nth_of_type_3n_2
PASS Remove #div3: #nth_of_type_3n
PASS Remove #div4: #only_of_type
FAIL Remove #div4: #first_of_type assert_equals: expected "rgb(128, 128, 128)" but got "rgb(255, 165, 0)"
PASS Remove #div4: #last_of_type
PASS Remove #div4: #nth_of_type_3n_1
PASS Remove #div4: #nth_of_type_3n_2
PASS Remove #div4: #nth_of_type_3n
PASS Remove #div5: #only_of_type
FAIL Remove #div5: #first_of_type assert_equals: expected "rgb(128, 128, 128)" but got "rgb(255, 165, 0)"
PASS Remove #div5: #last_of_type
PASS Remove #div5: #nth_of_type_3n_1
PASS Remove #div5: #nth_of_type_3n_2
PASS Remove #div5: #nth_of_type_3n
